arm32: Correctly compute R_ARM_GOT_PREL.
Fixes #550.

In the Arm32 [spec][spec], R_ARM_GOT_PREL is defined as follows:

R_ARM_GOT_PREL: GOT(S) + A – P

Where:
> GOT(S) is the address of the GOT entry for the symbol S.

If we look at the code a bit, notice that G is actually the _offset_
of S into the GOT, not the symbol's address in GOT. This patch makes
sure that the relocation is computed using the GOT address of the
symbol, thereby fixing the mentioned issue.

[spec]: https://github.com/ARM-software/abi-aa/blob/60a8eb8c55e999d74dac5e368fc9d7e36e38dda4/aaelf32/aaelf32.rst#5612relocation-types

Signed-off-by: Robert Bartlensky <bartlensky.robert@gmail.com>