Revert "[RISCV32] Fix R_RISCV_HI20 relocations"
This reverts commit e65cc7feea10792432fb89f78eeae76b4ef3c7c8 because
I don't think this is a correct fix.

We can't emit a R_RISCV_32 dynamic relocation for R_RISCV_HI20
relocation. If we do, the dynamic loader will override an entire 32
bit word with an address of a specified symbol. HI20 is expected to
set a high 20 bits as an instruction operand, so they are not
compatible.

https://github.com/rui314/mold/pull/668