[ELF][SPARC] Fix endianness issue
Bitfields are placed in an address-increasing order on both little
and big-endian hosts. Therefore, the following struct has the same
layout on both endiannesses.

  struct {
    uint32_t x : 8;
    uint32_t y : 16;
  }

The address of `x` is smaller than `y`, though you can't take an
address of a bitfield using the `&` operator in C.

However, if two bitfields are placed the same byte, they are placed in
a reverse order. In LE, bitfields are placed from LSB to MSB, while in
BE, they are placed from MSB to LSB. Therefore, members something like
this

  struct {
    uint32_t x : 4;
    uint32_t y : 4;
  }

needs to be swapped.