Add precompiled headers for CMake builds.
Since we're having lots of arch-generic permutations, headers seems to be a good
place where potential perf improvement are possible.

The feature is gated against CMake 3.16.0 or later, which is when it was
implemented.

Measured on 3990X, with this change it takes:
ninja  1166.22s user 34.45s system 5378% cpu 22.322 total
While before it took:
ninja  1776.81s user 51.30s system 7172% cpu 25.489 total

So a 10% improvement for workstation users and even more if you don't.

Signed-off-by: Tatsuyuki Ishi <ishitatsuyuki@gmail.com>