[Mach-O] Support `-flat_namespace` flag
This implements the minimum necessary to support the `-flat_namespace`
flag.

I have not added this flag to the `-help` text, because I'm not certain
this is faithful to ld64's behaviour under this flag. From `man ld`:

     -flat_namespace
             Alters how symbols are resolved at build time and runtime.  With -two_levelnamespace (the default), the linker only
             searches dylibs on the command line for symbols, and records in which dylib they were found.  With -flat_namespace, the
             linker searches all dylibs on the command line and all dylibs those original dylibs depend on.  The linker does not record
             which dylib an external symbol came from, so at runtime dyld again searches all images and uses the first definition it
             finds.  In addition, any undefines in loaded flat_namespace dylibs must be resolvable at build time.

However, I believe what is done here is enough to approximate what build
systems expect to happen when they pass `-flat_namespace` to the linker.

Closes #785.

Signed-off-by: Carlo Cabrera <30379873+carlocab@users.noreply.github.com>