[ELF] Re-implement --relocatable
This is an initial patch to reimplement --relocatable. Unlike the previous
implementation, the new one aims to be on par with GNU ld in terms of the
behavior. Previously, we didn't for example merge input sections at all.
Now we do it in the new implementation.

The previous implementation was slow. We also want to fix it in the new one.

The new --relocation feature is not complete yet; there are many rough edges.
But I think this patch is good enough to be submitted.

This work is funded by Mercury (https://mercury.com).