Create `mold-config.cmake`
With this commit, a config file for mold will be generated and
installed to `${CMAKE_INSTALL_LIBDIR}/cmake/mold/mold-config.cmake`.

The config file intends to help users to get started with mold easily.
It contains a function `target_use_mold`, which sets mold as the linker
for the target specified. For more information, see the comments in
`mold-config.cmake.in`.

Signed-off-by: Zhong Ruoyu <zhongruoyu@outlook.com>