Dockerfile: fix OpenSSL libdir
In the current version of Docker image, OpenSSL libraries are installed
to `/usr/local/lib64`. It is not present in the default dynamic linker
search path, and thus results in the following error when libcrypto is
not statically-linked:

  mold: error while loading shared libraries: libcrypto.so.3: cannot
  open shared object file: No such file or directory

Fix that by overriding `libdir`, as suggested in
https://github.com/openssl/openssl/blob/master/INSTALL.md#libdir. Also,
run `ldconfig` to update dynamic linker cache.

Signed-off-by: Ruoyu Zhong <zhongruoyu@outlook.com>