[ELF] Allow to initialize a global variable with an ifunc function pointer
This is a tricky case, but if you initialize a global variable with a
pointer to an ifunc function like this:

  int bar() __attribute__((ifunc("resolve_bar")));
  void *bar2 = bar;

then we need to emit two dynamic relocations, one for `bar()` and the other
for `bar2`. This results in calling `resolve_bar` twice at runtime, but
there's no other way to initialize the function and the variable correctly.

Fixes https://github.com/rui314/mold/issues/915