Do not bother to clear existing bits when applying relocations
We were inconsistent as to when to clear existing values when applying
relocations. mold was broken from the point of view that bits must be
cleared before applying relcoations. But at the same time, we had code
to clear bits for some types of relocations.

This patch removes all bit clear code for the sake of simplicity.