Fix search order in INPUT and GROUP of script
The scenario is from incremental cross compiling gcc.
libgcc_s.so is a ld script in my OS.
  GROUP ( libgcc_s.so.1 -lgcc )
The path of libgcc_s.so.1 is not absolute.
When current directory contains libgcc_s.so.1 of other ISA,
We will get target of other ISA and report "skipping incompatible file".

As ld manual shows:
If a sysroot prefix is not used then the linker will try to open
the file in the directory containing the linker script.
If it is not found the linker will then search the current directory.
If it is still not found the linker will search through the archive library search path.